+++ src/slide/index.ts
@@ -8,3 +8,4 @@ export * from './slide-body/slide-body-section.component';
 export * from './slide-footer/slide-footer.component';
 export * from './slide-container.component';
 export * from './slide.config';
+export * from './drawer/drawer-container.directive';